Handle time going backwards, and net consumption never going decreasing (#11)
* Add W_consumed and W_returned Sensors for solar

If a solar customer is not yet on _net metering_, then the kWh_net
Sensor will never decrease (i.e., meter does _not_ run
backwards). This means that kWh_returned will always be zero, even
though power _is_ flowing back to the grid (though the customer is not
being credited for it).

For visualizing in e.g. Home Assistant's Energy dashboard, we
therefore export the Watts flowing in either direction (to or from the
grid) so that integral helpers/sensors/entities can be used to show
the _actual_ energy flow to and from the grid (again, even though the
customer is not being credited (yet?) on the meter / bill).

* disambiguate `time()` if e.g. using the ESPHome `time` component

* handle time jumping backwards